crud with auth fix

+++ app/Http/Controllers/Api/UserController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use App\Http\Resources\UserResource;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -71,7 +72,13 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        $user->delete();
-        return response("", 204);
+        if(Auth::user()->id !== $user->id){
+            $user->delete();
+            return response("", 204);
+        }else{
+            $data = "You cannot delete your own account!";
+            return response($data, 403);
+        }
+        
     }
 }
